backend/requirements.txt
fastapi
uvicorn
nsepython
pandas
schedule
pydantic
python-dotenv
requests
